requirements.txt
aio-pika==6.6.1
aiohttp==3.6.2
aioredis==1.3.1
aiormq==3.2.2
python-engineio==4.0.0
python-socketio==5.0.3
websocket-client==0.53.0
requests==2.22.0
asyncio-nats-client==0.11.4
